+++ src/app/core/type/app.type.ts
@@ -1,7 +1,8 @@
 export enum VisibleFlag {
     None = 0,
     Game = 1 << 0,
-    Dialog = 1 << 1
+    Dialog = 1 << 1,
+    Browser = 1 << 2
 }
 
 export enum AppUpdateState {