+++ src/app/data/poe-prices/schema/item-price-prediction.ts
@@ -1,0 +1,8 @@
+export interface ItemPricePredictionResponse {
+    min: number;
+    max: number;
+    currency: string;
+    warning_msg: string;
+    error_msg: string;
+    pred_confidence_score: number
+}