+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.6.22 (2020-04-13)
+
+- add poeprices.info price prediction (#544)
+- fix multi dialogs not working
+- fix default range values applied on properties (#615)
+- fix apply range only on attack/ defense properties (#617)
+- fix poe overlay does not skip the task bar (#618)
+
 ## 0.6.21 (2020-04-12)
 
 - add focus game after closing all dialogs (#606)